+++ Src/FluentAssertions/Execution/Execute.cs
@@ -1,4 +1,6 @@
-﻿namespace FluentAssertions.Execution;
+﻿using FluentAssertions.Common;
+
+namespace FluentAssertions.Execution;
 
 /// <summary>
 /// Helper class for verifying a condition and/or throwing a test harness specific exception representing an assertion failure.
@@ -8,5 +10,12 @@ public static class Execute
     /// <summary>
     /// Gets an object that wraps and executes a conditional or unconditional assertion.
     /// </summary>
-    public static AssertionScope Assertion => AssertionScope.Current;
+    public static AssertionScope Assertion
+    {
+        get
+        {
+            Services.EnsureInitialized();
+            return AssertionScope.Current;
+        }
+    }
 }